+++ application/views/partials/_menubar.php
@@ -5,9 +5,7 @@
     </div>
     <ul class="nav navbar-nav">
       <li class="active"><a href=/"welcome">Home</a></li>
-      <li><a href="/info/catalog">Catalog</a></li>
-      <li><a href="/info/category">Categories</a></li>
-      <li><a href="/info/bundle">Bundles</a></li>
+      <li><a href="/catalog">Catalog</a></li>
       <li><a href="/about">About</a></li>
     </ul>
   </div>