+++ application/views/partials/_menubar.php
@@ -4,10 +4,9 @@
       <a class="navbar-brand" href="#">PoEEquips</a>
     </div>
     <ul class="nav navbar-nav">
-      <li class="active"><a href="#">Home</a></li>
-      <li><a href="#">Page 1</a></li>
-      <li><a href="#">Page 2</a></li>
-      <li><a href="#">Page 3</a></li>
+      <li class="active"><a href=/"welcome">Home</a></li>
+      <li><a href="/catalog">Catalog</a></li>
+      <li><a href="/about">About</a></li>
     </ul>
   </div>
 </nav>