+++ application/views/partials/_menubar.php
@@ -4,7 +4,7 @@
       <a class="navbar-brand" href="/welcome">PoEEquips</a>
     </div>
     <ul class="nav navbar-nav">
-      <li class="active"><a href="/welcome">Home</a></li>
+      <li><a href="/welcome">Home</a></li>
       <li><a href="/catalog">Catalog</a></li>
       <li><a href="/about">About</a></li>
     </ul>