+++ changelog.md
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 # Changelog
 
 ## Rules
@@ -8,6 +7,9 @@
 * Changelog Format: Read from newest to oldest
 
 ## Version 0.1
+### New components
+--------Feb 10, 2018--------
+* Catalog Controller and View - QL
 
 ### New components
 --------Feb 9, 2018--------