+++ changelog.md
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 # Changelog
 
 ## Rules
@@ -27,9 +28,19 @@
 * about.php controllers and about_message view - RT
 
 ### Updated components
---------Feb 9,2018
-* ..assets/css/default.css (updated css file)
-* ../view/Welcome_message (added invantory image and divs)
+--------Feb 10,2018-------
+* updated menu items (delete link to Json files and replace with catalog) -CL
+* added links to menu dropDownItems -CL
+* updated ../view/about_message (Changed the index function to make the page looks nicer) -CL
+--------Feb 9,2018--------
+* updated Presets/ Categories/ Accessories to have gloves category - QL
+* added selector functionality for presets on front page - QL
+* populated selector with presets functional - QL
+* made pictures dynamically fill equipment boxes when selecting presets - QL
+* fixed image sized to fit with inventory slot size - QL
+* implemented catalog, bundle, category view controllers - QL
+* ..assets/css/default.css (updated css file) - CL
+* ../view/Welcome_message (added invantory image and divs) - CL
 * ../views/template (moved the footer to footer.php) -CL
 * ../cotrollers/Welcome.php (load partial and welcome page) -CL
-* gitignore (allowed CSV files in git)
+* gitignore (allowed CSV files in git) -CL