+++ changelog.md
@@ -36,6 +36,10 @@
 --------Feb 10,2018-------
 * move the image css to css files - RT
 * updated logo pictures and web app title - RT
+* updated ../view/welcome_message (changed the dropdown menu items name) -CL
+* updated ../view/welcome_message (added stats display) -CL
+* updated ../view/partial/menubar (updated link of home and logo) -CL
+* updated catalog controller to correctly display Category - QL
 * updated menu items (delete link to Json files and replace with catalog) -CL
 * added links to menu dropDownItems -CL
 * updated ../view/about_message (Changed the index function to make the page looks nicer) -CL